Record how our store_buf.idx evolves.
We want to record the value of the store_buf.idx each time a write is
made to the store_buf.

For this purpose, create another buffer called the idx_history such
that idx_history.elements[] contains the evolution of store_buf.idx.

Hence, each time a thread writes to store_buf.elements[] at an index
i, it will update idx_history.elements[] with that index value.

Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>