Scenario where we need to sleep in Critical Section
We create a scenario where we need to sleep inside the
write-critical-section where the number of writers are greater than
the number of CPUs on our system.

Hence add an msleep(10) in the thread_fn() between the writes to
store_buffer and idx_history buffers.

Also bump up the MAXTHREADS to 5.

Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>