Guard the writes to buffers in thread_fn by a mutex
Since the critical section inside thread_fn consists of a msleep()
which can take the task out of the CPU and out of the scheduler
runqueue, we cannot use spinlock as it can deadlock the system.

Task 1         Task 2        Task 3          Task 4         Task 5
==========     ============  ===========     ==========     =========
(CPU 0)        (CPU 1)       (CPU 2)         (CPU 3)
spin_lock()    spin_lock()   spin_lock()     spin_lock()

gets it        busyloop      busyloop        busyloop

.              .             .               .
.              .             .               .
.              .             .               .
msleep()       .             .		     .
               .             .		     .
goes out       .             .		     .
of runqueue   busyloop       busyloop        busyloop	    scheduled
                                                            on (CPU 0)

							    spin_lock()
							    busyloop

Thus, Tasks 2,3,4,5 are all busy looping on CPUs 0,1,2,3 waiting for
the spin_lock to be released. However Task 1 which holds the spinlock
is out of the CPU, and will never get a new CPU to run, since none of
the Tasks 2,3,4,5 will yield their CPUs.

Thus the system is deadlocked:

Task 0 ===> waiting for CPUs ===> occupied by Tasks 2,3,4,5
  AA                                     ||
  ||                                     ||
  ||                                     ||
  ||                                     VV
 Held by <======================== Waiting for spin lock

To fix this, guard the critical sections with a mutex. This type of
lock will allows the holders to sleep within the critical sections.

Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>