Implement our own spinlock
We implement a spinlock using cmpxchg. The lock is a struct containing
a unsigned 32-bit element.

We designate the lock as being in LOCKED state if only the MSB of this
element is set.

We designate the lock as being in an UNLOCKEDx state if this 32bit
element is 0.

We implement lock() and unlock() functions which use cmpxchg to
atomically change the state from LOCKED to UNLOCKED and vice-versa.

Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>