Add our own read-write lock
Define our own read-write lock with a u32 lock_and_count element.

The MSB will be used to indicate whether some writer holds the lock.

The lower 31 bits will be used to denote the readers-count.  Thus a
reader is said to hold a lock iff the lower 31 bits are non-zero.

If the value of the lock_and_count element is zero, it implies that
the lock is unlocked (neither readers, nor writers holds the lock)

The writer can take the lock iff the lock is in an UNLOCKED state.

The reader can take the lock iff the lock is not held by a writer. The
reader indicates that it holds the lock by incrementing the count
field of the lock_and_count element.

The writer unlocks the lock by atomically clearing the LOCKED bit
(MSB)

The reader releases the lock by atomically decrementing the count
field of the lock_and_count element.

Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>